Move EasyMDE init to Vue component
